requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pandas>=2.2.0
openpyxl==3.1.2
sqlalchemy>=2.0.25
psycopg2-binary>=2.9
asyncpg>=0.29
python-dotenv>=1.0.0
alembic>=1.13.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4
pydantic-settings>=2.0.0

# Email
aiosmtplib>=3.0.0
jinja2>=3.1.0
